pyproject.toml
[project]
name = "wali"
version = "0.1.0"
description = "Wallpaper switcher"
authors = [{ name = "Keith Hughitt" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "haishoku>=1.1.8",
    "Pillow>=10.2.0",
]

[project.scripts]
wali = "wali.cli:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
